# doc/openapi.yaml
swagger: "2.0"
info:
  version: 9.7.0
  title: "@fastify/swagger"
...
paths:
...
  /api/supervisors/l402-readiness/:
    get:
      responses:
        "200":
          description: Default Response
          schema:
            type: object
            required:
              - domainId
              - architecture
              - webhookEndpoint
              - secretManagerPath
              - checklistApprovals
              - diagnostics
            properties:
              id:
                type: number
              domainId:
                type: number
              architecture:
                type: string
              webhookEndpoint:
                anyOf:
                  - type: string
                  - type: "null"
              secretManagerPath:
                anyOf:
                  - type: string
                  - type: "null"
              checklistApprovals:
                type: object
                additionalProperties:
                  type: string
              diagnostics:
                type: object
                required:
                  - hasLnbitsUrl
                  - hasLnbitsKey
                  - webhookConfigured
                properties:
                  hasLnbitsUrl:
                    type: boolean
                  hasLnbitsKey:
                    type: boolean
                  webhookConfigured:
                    type: boolean
        "400":
          description: Default Response
          schema:
            type: object
            required:
              - error
              - code
              - remediation
            properties:
              error:
                type: string
              code:
                type: string
              remediation:
                type: string
        "401":
          description: Default Response
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
    put:
      parameters:
        - name: body
          in: body
          schema:
            type: object
            properties:
              architecture:
                type: string
                enum:
                  - mock
                  - lnbits
              webhookEndpoint:
                type: string
              secretManagerPath:
                type: string
              checklistApprovals:
                type: object
                additionalProperties:
                  type: string
      responses:
        "200":
          description: Default Response
          schema:
            type: object
            required:
              - ok
            properties:
              ok:
                type: boolean
                enum:
                  - true
        "400":
          description: Default Response
          schema:
            type: object
            required:
              - error
              - code
              - remediation
            properties:
              error:
                type: string
              code:
                type: string
              remediation:
                type: string
        "401":
          description: Default Response
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string